NXP Kinetis: Added GNAT Ravenscar SFP runtime port for the FRDM-KL25Z board
This port is based on the FRDM-K64F board port, but modified to support
the Cortex-M0+ core of the Kinetis KL25Z microcontroller. In particular,
the context switch code had to be modified to use only arv6-m
instructions.

NOTE FOR USERS OF THIS LIBRARY:

Since the Kinetis KL25Z MCU has an ARM Cortex-M0+ core, the library needs
to be compiled using -mcpu=cortex-m0plus. However, for some reason, even
with specifying -mcpu=cortex-m0plus in the library's runtime.xml file,
the linker links the libgcc for armv7-m instead of the armv6-m libgcc.
This causes the wrong __aeabi_idiv to be linked. The armv7-m
__aeabi_idiv uses an 'it' instruction which does not exist for
Cortex-m0+ (armv6-m). This causes the code generated for division
expressions to trigger a hardware exception.

A workaround for this problem is to replace libgcc.a in
your <gnat-installation-path>/lib/gcc/arm-eabi/4.9.4 with the
armv6-m flavor of this library.

Signed-off-by: German Rivera <jgrivera67@gmail.com>